en local me funciona perfecto ya con todas las optimiaciones pero en remoto todavia hay problemas. se hizo una adicion al docker file a ver si si se puede solucionar el problema

+++ services/informe_mensual_data.py
@@ -55,6 +55,8 @@ def it_is_outdated():
 
 def scratch_initialization():
     os.makedirs(DATA_DIR_INF_MENSUAL, exist_ok=True)
+    if os.path.exists(VMENSUAL_FILEPATH):
+        return
     df = general.fetch_socrata_datosgov(get_query_vmensual(), DB_ALIAS_VOL_MAYORISTAS, 20)
     df.to_csv(VMENSUAL_FILEPATH, index=False)
